241105 16:16 5)손님 0.5초에 음식 하나씩 먹기.

+++ Assets/Scripts/Systems/GameManager.cs
@@ -36,7 +36,7 @@ public class GameManager : SingletonManager<GameManager>
 	public bool isCounterUp = false;
 
 	[Tooltip("초당 음식 먹는 속도")]
-	public int[] eatFoodSpeedPerSeceond = { 2, 2, 2, 2, 2, 2 };
+	public float[] eatFoodSpeedPerSeceond = { 2f, 2f, 2f, 2f, 2f, 2f };
 
 	public float villianTimer = 0;
 	public bool needNewVillianTimerSetting = true;
